Pass defaultAccount to not call getState()
This was the error when getState() was called in contractCall() in the nested token creation calls: "You may not call store.getState() while the reducer is executing. The reducer has already received the state as an argument. Pass it down from the top reducer instead of reading it from the store."