Fix condition to show MetaMask warning: must check window.web3
I hope this fixes it under all conditions. Tried out a few states, not sure I got them all though. This seems to have changed recently - the browser didn't use to make a connection to the RPC Server directly and as such relying on the availability of props.defaultAccount was enough of an indicator. Thanks to @Mozeps for pointing out this bug! :+1: